Add ir/ and dotations/ subdirs in parameters/
