+++ redux/reducers/parameters/base/dotations.ts
@@ -1,0 +1,6 @@
+// eslint-disable-next-line no-unused-vars
+import { DotationsState } from "../interfaces";
+
+export const BASE_DOTATIONS_DEFAULT_STATE: DotationsState = {};
+
+export const dotations = (state = BASE_DOTATIONS_DEFAULT_STATE) => state;
