+++ redux/reducers/parameters/index.ts
@@ -4,7 +4,7 @@ import { amendement } from "./amendement";
 import { base } from "./base";
 import { plf } from "./plf";
 
-export * from "./parameters-state";
+export * from "./interfaces";
 
 export const parameters = combineReducers({
   amendement,
