+++ redux/reducers/parameters/interfaces/dotations-state.ts
@@ -1,0 +1,1 @@
+export interface DotationsState {}
